+++ kopeme-junit3/src/test/java/de/dagere/kopeme/junit/exampletests/runner/CopyOfJUnitTimeoutTest.java
@@ -34,7 +34,7 @@ public class CopyOfJUnitTimeoutTest extends KoPeMeTestcase {
 	}
 
 	@Override
-	protected int getMaximalTime() {
+	protected long getMaximalTime() {
 		return 1000;
 	}
